+++ src/components/Create.js
@@ -1,11 +1,9 @@
-import React, { useState, useEffect } from "react";
+import React, { useState } from "react";
 
 import ButtonToolbar from 'react-bootstrap/ButtonToolbar';
-import db from "../utils/firebase.config";
 
 import FileUpload from './FileUpload';
 import Dropdown from "./Dropdown";
-import SimpleBarChart from './SimpleBarChart';
 
 
 const Create = () => {
@@ -14,28 +12,8 @@ const Create = () => {
     const [selectedGraph, setSelectedGraph] = useState("");
     const [selectedXAxes, setSelectedXAxes] = useState("");
     const [selectedYAxes, setSelectedYAxes] = useState("");
-    const [users, setUsers] = useState([]);
-
-    const fetchUsers = async() => {
-        const response = db.collection('users');
-        const data = await response.get();
-        data.docs.forEach(element => {
-            setUsers([...users, element.data()]);
-            console.log(element.data());
-        });
-    }
-
-
-    useEffect(() => {
-        fetchUsers();
-      }, [])
 
     if(jsonData) {
-
-        // users.map(user => {
-        //     return console.log(user.username)
-        // })
-
         return (
         <div className="header">
             <h1>Choose type of graph</h1>
