+++ src/home.js
@@ -1,4 +1,5 @@
-import React from "react";
+import React, { useEffect, useState } from "react";
+import db from "./utils/firebase.config";
 import Button from 'react-bootstrap/Button';
 import ButtonToolbar from 'react-bootstrap/ButtonToolbar';
 import { LinkContainer } from 'react-router-bootstrap';
@@ -8,7 +9,25 @@ import LogoutButton from './components/login/Logout';
 import User from './components/login/User';
 
 function Home() {
-  const { isAuthenticated } = useAuth0();
+  const { user, isAuthenticated } = useAuth0();
+  const [authedUser, setAuthedUser] = useState("");
+
+    const fetchUsers = async() => {
+    const response = db.collection('users');
+    const data = await response.get();
+    data.docs.forEach(element => {
+        // setUsers([...users, element.data()]);
+        if (user && element.data().email === user.email) {
+          setAuthedUser(element.data());
+          console.log(`From database: ${JSON.stringify(element.data())}`);
+        }
+    });
+  }
+
+  useEffect(() => {
+    fetchUsers();
+  }, [isAuthenticated])
+
     return (
       <div className="wrapper">
               {!isAuthenticated ? (
